Move bookmark creation logic to bookmark module
- all done in background still; logic unchanged
