+++ src/background.js
@@ -1,18 +1,14 @@
 import urlRegex from 'url-regex'
-import { dataURLToBlob } from 'blob-util'
 
 import 'src/activity-logger/background'
 import 'src/blacklist/background'
 import 'src/search/background'
+import 'src/bookmarks/background'
 import 'src/omnibar'
 import { installTimeStorageKey } from 'src/imports/background'
-import fetchPageData from 'src/page-analysis/background/fetch-page-data'
 import { generatePageDocId } from 'src/page-storage'
-import * as index from 'src/search/search-index'
-import db from 'src/pouchdb'
 import { generateVisitDocId } from 'src/activity-logger'
 import { generateBookmarkDocId } from 'src/bookmarks'
-import { transformToBookmarkDoc } from 'src/imports'
 
 export const dataConvertTimeKey = 'data-conversion-timestamp'
 export const OVERVIEW_URL = '/overview/overview.html'
@@ -21,47 +17,6 @@ export const OVERVIEW_URL = '/overview/overview.html'
 window.generatePageDocId = generatePageDocId
 window.generateVisitDocId = generateVisitDocId
 window.generateBookmarkDocId = generateBookmarkDocId
-
-// TODO: Move all this bookmarks logic away
-async function getAttachments(pageData) {
-    const favIconBlob = await dataURLToBlob(pageData.favIconURI)
-
-    const _attachments = {
-        'fav-icon-blob': {
-            content_type: favIconBlob.type,
-            data: favIconBlob,
-        },
-    }
-
-    return _attachments
-}
-
-async function createNewPageForBookmark(bookmarkInfo) {
-    let pageDoc = {
-        url: bookmarkInfo.url,
-        title: bookmarkInfo.title,
-        _id: generatePageDocId({ url: bookmarkInfo.url }),
-    }
-
-    try {
-        const fetchPageDataOptions = {
-            includePageContent: true,
-            includeFavIcon: true,
-        }
-        const pageData = await fetchPageData({ url: bookmarkInfo.url, opts: fetchPageDataOptions })
-        pageDoc = {
-            ...pageDoc,
-            _attachments: await getAttachments(pageData),
-            content: pageData.content,
-        }
-    } catch (err) {
-        console.error("Error occurred while fetching page data: " + err.toString())
-    } finally {
-        const bookmarkDoc = transformToBookmarkDoc(pageDoc)(bookmarkInfo)
-        index.addPageConcurrent({ pageDoc, bookmarkDocs: [bookmarkDoc] })
-        db.bulkDocs([bookmarkDoc, pageDoc])
-    }
-}
 
 async function openOverview() {
     const [currentTab] = await browser.tabs.query({ active: true })
@@ -73,23 +28,6 @@ async function openOverview() {
         browser.tabs.update({ url: OVERVIEW_URL })
     }
 }
-
-browser.bookmarks.onCreated.addListener(async (id, bookmarkInfo) => {
-    try {
-        // Attempt to resolve existing page doc to connect new bookmark to
-        const pageDoc = await db.get(generatePageDocId({ url: bookmarkInfo.url }))
-        const bookmarkDoc = transformToBookmarkDoc(pageDoc)(bookmarkInfo)
-        index.addBookmark(bookmarkDoc)
-        db.bulkDocs([bookmarkDoc, pageDoc])
-    } catch (error) {
-        // Create new page + bookmark if existing page not found
-        if (error.status === 404) {
-            createNewPageForBookmark(bookmarkInfo)
-        } else {
-            console.error(error) // Can't handle other errors; log to background script console
-        }
-    }
-})
 
 browser.commands.onCommand.addListener(command => {
     if (command === 'openOverview') {
